replaces GenerateSlashCommand with SlackSlashCommandGenerator

+++ src/SlackSlashCommandGenerator.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Willdone\SlackSlashCommand;
+namespace Wzulfikar\SlackSlashCommand;
 
 use Illuminate\Console\Command;
 
@@ -11,7 +11,7 @@ class SlackSlashCommandGenerator extends Command
      *
      * @var string
      */
-    protected $signature = 'make:slack_slash_command {className?}';
+    protected $signature = 'make:slack-slash-command {className?}';
 
     /**
      * The console command description.
@@ -19,8 +19,6 @@ class SlackSlashCommandGenerator extends Command
      * @var string
      */
     protected $description = 'Generate class implementation of SlackSlashCommandInterface.';
-
-    public $config;
 
     /**
      * Create a new command instance.
@@ -30,7 +28,6 @@ class SlackSlashCommandGenerator extends Command
     public function __construct()
     {
         parent::__construct();
-        $this->config  = (object) require (__DIR__.'/slack_slash_command.php');
     }
 
     /**
@@ -41,18 +38,35 @@ class SlackSlashCommandGenerator extends Command
     public function handle()
     {
         // prepare class name from given command
-        $command           = $this->argument('className') ?: $this->ask('What is the command name?');
-        $commandName       = str_replace('/', '', $command);
-        $commandsNamespace = $this->config->commandsNamespace;
+        $command      = $this->argument('className') ?: $this->ask('What is the command name?');
+        $commandName  = str_replace('/', '', $command);
+        $commandsDir  = config('slack_slash_command.commands_dir');
 
-        $className         = sprintf('Slash%s',ucfirst($commandName));
-        $classPath         = app_path($commandsNamespace . '/' . $className . '.php');
+        if(! $commandsDir ){
+            $this->error("Cannot fetch commands dir from config file. Have you done `php artisan vendor:publish`?");
+            exit();
+        }
+
+        if (!file_exists($commandsDir)) {
+            mkdir( $commandsDir, 0777);
+            $this->info("The directory $commandsDir was successfully created.");
+        }
+
+        $className    = sprintf('Slash%s',ucfirst($commandName));
+        $classPath    = $commandsDir . '/' . $className . '.php';
+
+        if (file_exists($classPath)) {
+            if (! $this->confirm('File '.$classPath.' already exists. Do you wish to continue & overwrite current file? [y|N]')){
+                $this->info('See ya ;)');
+                exit();
+            }
+        }
 
         // prepare template
-        $template = file_get_contents(app_path('Willdone/SlackSlashCommand/SlackSlashCommandTemplate.php'));
+        $template = file_get_contents(__DIR__ . ('/SlackSlashCommandTemplate.php'));
         $template = str_replace('{$className}', $className, $template);
         $template = str_replace('{$commandName}', $commandName, $template);
-        $template = str_replace('{$commandsNamespace}', $commandName, $template);
+        $template = str_replace('{$commandsNamespace}', str_replace('/', '\\', $commandsDir), $template);
 
         // create class definition from template
         file_put_contents($classPath, $template);
